+++ impl/model/src/main/java/io/serverlessworkflow/impl/model/jackson/JacksonModelCollection.java
