+++ impl/model/src/main/java/io/serverlessworkflow/impl/model/jackson/JacksonModelFactory.java
@@ -27,8 +27,8 @@ import io.cloudevents.CloudEventData;
 import io.serverlessworkflow.impl.WorkflowModel;
 import io.serverlessworkflow.impl.WorkflowModelCollection;
 import io.serverlessworkflow.impl.WorkflowModelFactory;
+import io.serverlessworkflow.impl.jackson.JacksonCloudEventUtils;
 import io.serverlessworkflow.impl.jackson.JsonUtils;
-import io.serverlessworkflow.impl.jackson.events.JacksonCloudEventUtils;
 import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.Map;
