+++ impl/test/pom.xml
@@ -31,10 +31,6 @@
         </dependency>
         <dependency>
             <groupId>io.serverlessworkflow</groupId>
-            <artifactId>serverlessworkflow-impl-jq</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.serverlessworkflow</groupId>
             <artifactId>serverlessworkflow-impl-jackson-jwt</artifactId>
         </dependency>
         <dependency>
